+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Enabled validation against applying `Skip` attributes to struct fields.
   * Fixed an issue when empty documentation comment was generated in C++ instead of no documentation comment.
   * Updated HTML formatting of generated JavaDoc to be more consistent.
+  * Fixed Java and C++ compilation issues for struct fields of `Locale` type.
 
 ## 8.4.5
 Release date: 2020-10-08